+++ osidb/signals.py
@@ -66,6 +66,9 @@ def update_major_incident_start_dt(flaw: Flaw) -> None:
     is_major_incident = flaw.major_incident_state in {
         Flaw.FlawMajorIncident.APPROVED,
         Flaw.FlawMajorIncident.CISA_APPROVED,
+        # Flaw.FlawMajorIncident.MINOR is not
+        # included as it has no impact on the SLA
+        Flaw.FlawMajorIncident.ZERO_DAY,
     }
     if is_major_incident and flaw.major_incident_start_dt is None:
         flaw.major_incident_start_dt = timezone.now()